+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `table.DefaultTxControl()` transaction control creator with serializable read-write isolation mode and auto-commit
+* Fixed passing nil query parameters
 * Fixed locking of cluster during call `cluster.Get`
 
 ## v3.19.1